requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
dnspython==2.7.0
greenlet==3.1.1
idna==3.10
playwright==1.51.0
pyee==12.1.1
pymongo==4.11.3
PyYAML==6.0.2
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.3.0
